CLIN-452: Fix itemCount
