+++ src/components/table/ItemsCount/index.module.scss
@@ -1,0 +1,3 @@
+.itemCount {
+    margin-bottom: 16px;
+}
