+++ src/locales/fr.js
@@ -1147,6 +1147,8 @@ const fr = {
   'screen.variant.entity.frequencyTab.analysis': "Analyse",
   'screen.variant.entity.frequencyTab.card.title': "Nombre de patients avec le variant {variant}",
   'screen.variant.entity.frequencyTab.RQDM.title': "Réseau québecois de diagnostic moléculaire",
+  'component.table.itemcount.results': 'Résultats',
+  'component.table.itemcount.of': 'sur',
 };
 
 export default fr;
