+++ src/views/screens/search/table/table.module.scss
@@ -1,3 +1,1 @@
-.headerInfo {
-  margin-bottom: 12px;
-}
+
