+++ src/views/screens/variant/VariantTableContainer.module.scss
@@ -38,10 +38,3 @@
     }
   }
 }
-
-.tabletotalTitle {
-  font-size: 14px;
-  font-weight: 400;
-  color: $gray-8;
-  margin-bottom: 16px;
-}
